requirements.txt
Flask==3.0.0
psutil==5.9.6
requests==2.31.0
python-dotenv==1.0.0

